Update TEN HD picon links
When TEN HD started broadcasting, it used service type 1 (generic
digital TV). When it launched, it changed the service type to 19
(MPEG4 HD TV), and the picon links generated from data collected
before the launch need to be updated to reflect this.

This affects regions Adelaide, Brisbane, Geelong, Melbourne, Perth,
Sunshine Coast and Sydney, but only changes 5 links for each picon
type (screen or lcd), because the TEN HD servicerefs tor the pairs
(Brisbane, Sunshine Coast) and (Melbourne, Geelong) are the same.