Updated picons for NITV and RACING.COM
Use a drop shadow as a background for the on-screen NITV and
RACING.COM picons and for the T4 front panel NITV picon.

This replaces a solid white background for the NITV picons and gives
better contrast for the on-screen RACING.COM picon.

This commit also changes the name for the RACING.COM picon from
racingcom.png to racingcom_lw.png to indicate the origin of the
picon (LogoWikia).