Change ABC2/KIDS logo to ABC COMEDY/KIDS logo
ABC LCN 22 changed from ABC2 + ABC KIDS to ABC COMEDY + ABC KIDS
on 4/12/2017.

Updated the logos for the channel to replace the ABC2 logo with the
ABC COMEDY logo in the combined picon.

The picon filename base is "abccomedy", reflecting the change in
the service\ name from ABC2 / KIDS to just ABC COMEDY.

The servicerefs for the channel don't appear to have changed.