Remove trailing spaces in picondefs
Some scan data has a trailing space in the text part of some
<e2servicename/> elements.

This causes non-functional differences in the picondefs file, and
these differences can vary and cause spurious changes in the picondefs
file.

The changes to the picondefs file affects only the end-of-line
comments with the original service name, and this does not affect
the actual picon setup.

The script generating the picondefs file has been modified to
strip trailing (and leading) space from service names before
processing them to create picon names.