Change ABC picon setup for ABC Comedy/ABC TVPlus rebadge
The picon file name has changed from abccomedy_*.png to
abckidstvplus_*.png, and the picon links have changed to reflect
that.

New scan files for Melbourne and Canberra were used, all other
changes for the rebadge were handled by renamings in the picon setup
scripts.

The new scan files for Canberra and Melbourne changed the service
type for ABC News Radio from 0x2 (dRadio) to 0xA (dRadioAvc). This
may be a national change, and might need a tweak in the setup
scripts.