New picons for ABC, ABC HD and ABC News
Update picons for ABC & ABC HD to the new Jan 2021 picons.

Change the ABC News screen picon to the stacked version for a better
fit.  This isn't a new picon style. The T4 front panel oicon is
unchanged.

This also removes a few remappings where picon links with service type
0xA were being duplicated with service type 0x1.