Add specific 7two HD picons
The previous picons for 7two HD were copies of the 7two picons.
This is apparently in line with the broacdaster use.

Replace those picons with a specific "HD" tagged 7two HD picons in
a similar style to the 7mate HD picons, to differentiate the picons
for 7two and 7two HD.