Add 7bravo in metro regions and 7HD and 7twoHD fixups for Qld regionals and Perth
Add new 7bravo service.

In metro areas and some regional Queensland areas, this changes the
picon for openshop's serviceref to the 7bravo picon.

In other regional areas, this changes the picon for ishoptv's
serviceref to the 7bravo picon, and outs the openshop picon on its
new serviceref.

Remove now-unused picon for openshop.

In picondefs, there are some non-functional incidental changes because
Orange is using new scan data for the changes.

Change picons for 7two to 7twoHD and 7mate to blank for Qld regionals
& Perth.

Change picons for 7 to 7HD for Qld regionals.

Change servicetype for TVSN Rockhampton to MPEG2 (from new scan
data).

Other incidental and non-functional changes for Rockhampton in
picondefs.